+++ packages/wonder-blocks-tooltip/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/wonder-blocks-tooltip
 
+## 4.1.18
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-typography@4.0.0
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-core@12.3.0
+    - @khanacademy/wonder-blocks-layout@3.1.15
+    - @khanacademy/wonder-blocks-modal@7.1.18
+
 ## 4.1.17
 
 ### Patch Changes